refactor: rename util folder to utils
