+++ src/extension/backgroundscript/utils/format.ts
